+++ examples/react/farm.config.js
@@ -1,24 +1,15 @@
 import { defineConfig } from '@farmfe/core';
-import path from 'path';
 
 export default defineConfig(() => {
-  // console.log(__dirname);
-  // console.log(__filename);
-  // console.log(__dirname);
-
   return {
-    // root: path.resolve(process.cwd(), './html'),
     compilation: {
       sourcemap: true,
       persistentCache: false,
       presetEnv: false,
       progress: false,
-      output: {
-        publicPath: '/dist/'
-      },
-      input: {
-        index: './index.html'
-      }
+      // output: {
+      //   publicPath: '/dist/'
+      // },
     },
     server: {
       port: 6532,