+++ packages/core/src/config/constants.ts
@@ -1,7 +1,10 @@
 export const DEFAULT_CONFIG_NAMES = [
   'farm.config.ts',
   'farm.config.js',
-  'farm.config.mjs'
+  'farm.config.cjs',
+  'farm.config.mjs',
+  'farm.config.cts',
+  'farm.config.mts'
 ];
 
 export const FARM_DEFAULT_NAMESPACE = 'FARM_DEFAULT_NAMESPACE';