+++ packages/runtime-plugin-import-meta/src/index.ts
@@ -11,7 +11,6 @@ export default (<Plugin>{
   moduleCreated(module) {
     module.meta.env = {
       ...((FARM_PROCESS_ENV) ?? {}),
-      mode: process.env.NODE_ENV,
       dev: process.env.NODE_ENV === 'development',
       prod: process.env.NODE_ENV === 'production'
     };