+++ packages/core/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @farmfe/core
 
+## 1.2.6
+
+### Patch Changes
+
+- a264f6c4: support mts/cts ext and fix traceDependencies use directory
+- 0915328e: support new URL with import.meta.url
+
+## 1.2.5
+
+### Patch Changes
+
+- 49523be7: bump dotenv version
+- 53f5115e: fix: publicDir find html resource middleware error
+- Updated dependencies [49523be7]
+  - @farmfe/runtime-plugin-import-meta@0.2.1
+
 ## 1.2.4
 
 ### Patch Changes